rename ios and android run files
